cleaned unused code, handle errors
