+++ bot.js
@@ -7,10 +7,9 @@ const { handleMessage } = require("./commands");
 
 class Bot {
     constructor() {
-        this.config = require("../config.json");
         this.client = new Client({
             authStrategy: new LocalAuth(),
-            puppeteer: {headless: true, args: ['--no-sandbox', '--disable-setuid-sandbox']}
+            puppeteer: {headless: true}
         });
 
         this.client.on('loading_screen', (percent, message) => {
@@ -38,8 +37,6 @@ class Bot {
         });
 
         this.client.on('message_create', async(message) => {
-            let logger;
-            let authorInfo;
             const author = await message.getContact();
             const chat = await message.getChat();
 
@@ -48,33 +45,14 @@ class Bot {
             console.log(message.body);
             console.log(chat.id.user);
             console.log("######################");
-            if (!this.config.blackList.includes(chat.id.user)) {
-                console.log("CHAT HABILITADO");
-                await handleMessage(message, chat, author,);
-            }
+            await handleMessage(message, chat, author);
         });
 
         this.client.initialize();
     }
-
-    addGroupBlackList(id) {
-        if (!this.config.blackList.includes(id)) {
-            this.config.blackList.push(id);
-        } else {
-            throw {Error: "Duplicated id"};
-        }
-    }
-
-    removeGroupBlackList(id) {
-        const index = this.config.blackList.indexOf(id);
-        if (index !== -1) {
-            this.config.blackList.splice(index, 1);
-        }
-    }
-
 }
 
-module.exports = { Bot };
+const bot = new Bot();
