+++ llm.js
@@ -11,9 +11,7 @@ const completion = async (message) => {
     const response = await openai.createChatCompletion({
         model: "gpt-3.5-turbo",
         messages: [{role: "user", content:  message}],
-        max_tokens: 1000
     });
-
     return response.data.choices[0].message.content;
 }
 
@@ -21,7 +19,7 @@ async function transcribe() {
     const resp = await openai.createTranscription(
         fs.createReadStream("temp/audio.mp3"),
         "whisper-1", undefined, undefined, undefined,
-        "es"
+        undefined
     );
     return resp.data.text;
 }
