+++ package.json
@@ -1,23 +1,22 @@
 {
-  "name": "untitled22",
+  "name": "untitled25",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
+  "main": "bot.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "author": "",
-  "license": "ISC",
   "dependencies": {
-    "express": "^4.18.2",
     "fluent-ffmpeg": "^2.1.2",
-    "node-fetch": "^2.6.9",
     "openai": "^3.2.1",
+    "puppeteer": "^20.1.2",
     "qr-image": "^3.2.0",
     "qrcode": "^1.5.3",
     "qrcode-terminal": "^0.12.0",
     "unidecode": "^0.1.8",
     "uuid": "^9.0.0",
     "whatsapp-web.js": "^1.19.5"
-  }
+  },
+  "author": "",
+  "license": "ISC"
 }
